Game: created separate file for draw functions and fixed bug where score overdraws itself
