+++ frontend/src/components/gamewindow/Gamewindow.tsx
@@ -1,14 +1,16 @@
 import React, { useRef, useEffect, useState } from "react";
-import properties, {
-  IBall,
-  IGame,
-  ballSpawn,
-  gameSpawn,
-  IPaddle,
-} from "./properties";
+import properties, { IGame, ballSpawn, gameSpawn, IPaddle } from "./properties";
 import Button from "../button";
 import Centerdiv from "../centerdiv";
 import Gamecanvas from "../gamecanvas";
+import {
+  drawBackground,
+  drawBall,
+  drawBothPaddles,
+  drawText,
+} from "./drawFunctions";
+
+const BACKEND: string = `http://${window.location.hostname}:${4000}`;
 
 interface IKeyState {
   up: boolean;
@@ -19,75 +21,21 @@ const fetchGameState = async (
   gameId: number,
   localPaddle: IPaddle
 ): Promise<IGame> => {
-  const response = await fetch(
-    `http://localhost:4000/games/gamestate/${gameId}`,
-    {
-      method: "POST",
-      headers: {
-        "Content-Type": "application/json",
-      },
-      body: JSON.stringify(localPaddle),
-    }
-  );
+  const response = await fetch(`${BACKEND}/games/gamestate/${gameId}`, {
+    method: "POST",
+    headers: {
+      "Content-Type": "application/json",
+    },
+    body: JSON.stringify(localPaddle),
+  });
   const ball: IGame = await response.json();
   return ball;
 };
 
 const stopAllGames = async (): Promise<void> => {
-  await fetch(`http://localhost:4000/games/stopall`, {
+  await fetch(`${BACKEND}/games/stopall`, {
     method: "POST",
   });
-};
-
-const drawPaddle = (
-  context: CanvasRenderingContext2D,
-  side: string,
-  height: number
-): void => {
-  const paddleHeight: number = Math.floor(
-    (properties.window.height * properties.paddle.height) / 100
-  );
-  const paddleWidth: number = Math.floor(
-    (properties.window.width * properties.paddle.width) / 100
-  );
-  if (side === "left") {
-    context.fillStyle = properties.window.color;
-    context.clearRect(0, 0, paddleWidth, properties.window.height);
-    context.fillStyle = properties.paddle.color;
-    context.fillRect(0, height - paddleHeight / 2, paddleWidth, paddleHeight);
-  } else if (side === "right") {
-    context.fillStyle = properties.window.color;
-    context.clearRect(
-      properties.window.width - paddleWidth,
-      0,
-      paddleWidth,
-      properties.window.height
-    );
-    context.fillStyle = properties.paddle.color;
-    context.fillRect(
-      properties.window.width - paddleWidth,
-      height - paddleHeight / 2,
-      paddleWidth,
-      paddleHeight
-    );
-  }
-};
-
-const drawBall = (
-  context: CanvasRenderingContext2D,
-  ball: IBall,
-  old_ball: IBall
-): void => {
-  context.clearRect(
-    old_ball.x - properties.ballProperties.radius * 2,
-    old_ball.y - properties.ballProperties.radius * 2,
-    properties.ballProperties.radius * 4,
-    properties.ballProperties.radius * 4
-  );
-  context.fillStyle = properties.ballProperties.color;
-  context.beginPath();
-  context.arc(ball.x, ball.y, properties.ballProperties.radius, 0, 2 * Math.PI);
-  context.fill();
 };
 
 const GameWindow: React.FC = () => {
@@ -132,14 +80,14 @@ const GameWindow: React.FC = () => {
   };
 
   const spawnGame = async (): Promise<void> => {
-    const incomingGameId = await fetch(`http://localhost:4000/games/start`, {
+    const incomingGameId = await fetch(`${BACKEND}/games/start`, {
       method: "POST",
     });
     gameId.current = await incomingGameId.json();
   };
 
   const stop = async (): Promise<void> => {
-    await fetch(`http://localhost:4000/games/stop/${gameId.current}`, {
+    await fetch(`${BACKEND}/games/stop/${gameId.current}`, {
       method: "POST",
     });
   };
@@ -153,27 +101,10 @@ const GameWindow: React.FC = () => {
   };
 
   useEffect(() => {
-    const canvas: HTMLCanvasElement = backgroundRef.current;
-    canvas.focus();
-    const context: CanvasRenderingContext2D =
-      backgroundRef.current.getContext("2d");
-    context.fillStyle = properties.window.color;
-    context.fillRect(0, 0, properties.window.width, properties.window.height);
-    context.fillStyle = properties.paddle.color;
-    const squareSize: number = properties.window.height / 32;
-    let currenSquarePos: number = squareSize;
-    while (currenSquarePos < properties.window.height) {
-      context.fillRect(
-        properties.window.width / 2 - squareSize / 2,
-        currenSquarePos - squareSize / 2,
-        squareSize,
-        squareSize
-      );
-      currenSquarePos += 2 * squareSize;
-    }
+    drawBackground(backgroundRef.current.getContext("2d"));
     window.addEventListener(
       "keydown",
-      function (e) {
+      (e) => {
         if (e.key === "ArrowUp") keyState.current.up = true;
         if (e.key === "ArrowDown") keyState.current.down = true;
       },
@@ -182,7 +113,7 @@ const GameWindow: React.FC = () => {
 
     window.addEventListener(
       "keyup",
-      function (e) {
+      (e) => {
         if (e.key === "ArrowUp") keyState.current.up = false;
         if (e.key === "ArrowDown") keyState.current.down = false;
       },
@@ -197,37 +128,18 @@ const GameWindow: React.FC = () => {
   }, []); // eslint-disable-line react-hooks/exhaustive-deps
 
   useEffect(() => {
-    const canvas: HTMLCanvasElement = paddleRef.current;
-    canvas.focus();
-    const context: CanvasRenderingContext2D =
-      paddleRef.current.getContext("2d");
-
-    drawPaddle(
-      context,
-      gameState.current.left.side,
-      gameState.current.left.height
-    );
-    drawPaddle(
-      context,
-      gameState.current.right.side,
+    drawBothPaddles(
+      paddleRef.current.getContext("2d"),
+      gameState.current.left.height,
       gameState.current.right.height
     );
   }, [gameState.current.left.height, gameState.current.right.height]);
 
   useEffect(() => {
-    const context: CanvasRenderingContext2D = scoreRef.current.getContext("2d");
-    const fontSize: number = properties.window.height / 6;
-    context.font = `${fontSize}px Arial`;
-    context.fillStyle = "white";
-    context.fillText(
-      `${gameState.current.pointsLeft}`,
-      properties.window.width / 3,
-      properties.window.height / 6
-    );
-    context.fillText(
-      `${gameState.current.pointsRight}`,
-      (properties.window.width * 2) / 3,
-      properties.window.height / 6
+    drawText(
+      scoreRef.current.getContext("2d"),
+      gameState.current.pointsLeft,
+      gameState.current.pointsRight
     );
   }, [gameState.current.pointsLeft, gameState.current.pointsRight]);
